+++ tests/dummy/app/Image.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace DummyApp;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\MorphTo;
+
+class Image extends Model
+{
+
+    /**
+     * @return MorphTo
+     */
+    public function imageable()
+    {
+        return $this->morphTo();
+    }
+}